requirements.txt
PyMuPDF==1.23.8
sentence-transformers==2.5.1
chromadb==0.4.22
openai==1.12.0
streamlit==1.33.0
python-dotenv==1.0.1
pypdf==4.1.0
numpy>=1.24.0
pandas>=2.0.0
python-multipart>=0.0.6
